Config production environment to upload files
To this feature is used Cloudinary due to its free
layer. It's necessary to create an account

To upload files Strapi uses something that they
call "providers". The upload provider is configurated
in the directory; extentions/upload/config/settings.js.
If the path doesn't exist then we have to create it

We use Cloudinary because Strapi has an example with
this provider

Upload provider documentation:
- https://strapi.io/documentation/3.0.0-beta.x/plugins/upload.html#configuration-per-envrionment

Cloudinary provider example:
- https://github.com/strapi/strapi-starter-blog/blob/master/extensions/upload/config/settings.js

Strapi providers list to upload files:
- https://www.npmjs.com/search?q=strapi-provider-upload-

Changes:
  - Adding strapi-provider-upload-cloudinary npm package
  - Adding env variables to config Cloudinary provider
  - Adding setting.js file to configure the provider
    to upload files in Cloudinary